LPS-79878 Move and rename classes.
