+++ modules/apps/web-experience/journal/journal-service/src/main/java/com/liferay/journal/internal/upgrade/v1_1_2/UpgradeJournalServiceVerify.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.internal.verify;
+package com.liferay.journal.internal.upgrade.v1_1_2;
 
 import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.asset.kernel.service.AssetEntryLocalService;
@@ -48,6 +48,7 @@ import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.LoggingTimer;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.verify.model.VerifiableUUIDModel;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.kernel.xml.Document;
 import com.liferay.portal.kernel.xml.Element;
@@ -66,9 +67,9 @@ import javax.portlet.PortletPreferences;
  * @author Alexander Chow
  * @author Shinn Lok
  */
-public class JournalServiceVerifyProcess extends UpgradeProcess {
+public class UpgradeJournalServiceVerify extends UpgradeProcess {
 
-	public JournalServiceVerifyProcess(
+	public UpgradeJournalServiceVerify(
 		AssetEntryLocalService assetEntryLocalService,
 		JournalArticleLocalService journalArticleLocalService,
 		JournalArticleResourceLocalService journalArticleResourceLocalService,
@@ -597,7 +598,7 @@ public class JournalServiceVerifyProcess extends UpgradeProcess {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		JournalServiceVerifyProcess.class);
+		UpgradeJournalServiceVerify.class);
 
 	private final AssetEntryLocalService _assetEntryLocalService;
 	private final JournalArticleLocalService _journalArticleLocalService;
@@ -609,5 +610,35 @@ public class JournalServiceVerifyProcess extends UpgradeProcess {
 	private final Portal _portal;
 	private final ResourceLocalService _resourceLocalService;
 	private final SystemEventLocalService _systemEventLocalService;
+
+	private static class JournalArticleResourceVerifiableModel
+		implements VerifiableUUIDModel {
+
+		@Override
+		public String getPrimaryKeyColumnName() {
+			return "resourcePrimKey";
+		}
+
+		@Override
+		public String getTableName() {
+			return "JournalArticleResource";
+		}
+
+	}
+
+	private static class JournalFeedVerifiableModel
+		implements VerifiableUUIDModel {
+
+		@Override
+		public String getPrimaryKeyColumnName() {
+			return "id_";
+		}
+
+		@Override
+		public String getTableName() {
+			return "JournalFeed";
+		}
+
+	}
 
 }
